requirements.txt
torch==2.5.1
torchvision==0.20.1
pillow==11.3.0
pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
seaborn==0.13.2
kaggle==1.7.4.5
jupyter==1.1.1